add status check for long time async resource

+++ restapi/resource_api_object.go
@@ -106,14 +106,33 @@ func resourceRestAPI() *schema.Resource {
 				Description: "Whether to emit verbose debug output while working with the API object on the server.",
 				Optional:    true,
 			},
-			"create_ready_key": {
+			"status_key": {
 				Type:        schema.TypeString,
 				Description: "The key to observe during resource creation. As long as its value is not equal to `create_ready_value` the resource is considered as pending. Similar to other configurable keys, the value may be in the format of 'field/field/field' to search for data deeper in the returned object.",
 				Optional:    true,
 			},
-			"create_ready_value": {
+			"status_ready_value": {
+				Type:        schema.TypeList,
+				Elem:        &schema.Schema{Type: schema.TypeString},
+				Description: "The value at `status_ready_value` indicating that a resource has been successfully created.",
+				Optional:    true,
+			},
+			"status_error_value": {
+				Type:        schema.TypeList,
+				Elem:        &schema.Schema{Type: schema.TypeString},
+				Description: "The value at `status_error_value` indicating that a resource failed created.",
+				Optional:    true,
+			},
+			"update_policy": {
 				Type:        schema.TypeString,
-				Description: "The value at `create_ready_key` indicating that a resource has been successfully created.",
+				Description: "The value at `update_policy` indicating that a resource update policy `none` `update`,`create`,`recreate`.",
+				Default:     "update",
+				Optional:    true,
+			},
+			"delete_policy": {
+				Type:        schema.TypeString,
+				Default:     "delete",
+				Description: "The value at `delete_policy`: `delete` or `orphan` .",
 				Optional:    true,
 			},
 			"read_search": {
@@ -224,33 +243,11 @@ func resourceRestAPICreate(d *schema.ResourceData, meta interface{}) error {
 
 	err = obj.createObject()
 	if err != nil {
-		return nil
+		return err
 	}
 
-	err = resource.Retry(d.Timeout(schema.TimeoutCreate), func() *resource.RetryError {
-		if obj.createReadyKey == "" || obj.createReadyValue == "" {
-			return nil
-		}
+	err = waitStatus(d, meta, obj)
 
-		err = obj.readObject()
-		if err != nil {
-			return resource.NonRetryableError(err)
-		} else if obj.id == "" {
-			return resource.NonRetryableError(fmt.Errorf("cannot evaluate readiness unless the ID has been set"))
-		}
-
-		readyValue, err := GetObjectAtKey(obj.apiData, obj.createReadyKey, obj.debug)
-		if err != nil {
-			return resource.NonRetryableError(err)
-		}
-
-		if fmt.Sprint(readyValue) == obj.createReadyValue {
-			/* Resource is ready and we can exit the retry loop */
-			return nil
-		} else {
-			return resource.RetryableError(fmt.Errorf("resource not yet ready - current value: %s", readyValue))
-		}
-	})
 	if err == nil {
 		/* Setting terraform ID tells terraform the object was created or it exists */
 		d.SetId(obj.id)
@@ -282,13 +279,52 @@ func resourceRestAPIRead(d *schema.ResourceData, meta interface{}) error {
 	}
 	return err
 }
+func waitStatus(d *schema.ResourceData, meta interface{}, obj *APIObject) error {
+	err := resource.Retry(d.Timeout(schema.TimeoutCreate), func() *resource.RetryError {
+		if obj.statusKey == "" || len(obj.statusReadyValue) == 0 {
+			return nil
+		}
+		err := obj.readObject()
+		if err != nil {
+			return resource.NonRetryableError(err)
+		} else if obj.id == "" {
+			return resource.NonRetryableError(fmt.Errorf("cannot evaluate readiness unless the ID has been set"))
+		}
 
+		readyValue, err := GetObjectAtKey(obj.apiData, obj.statusKey, obj.debug)
+		if err != nil {
+			return resource.NonRetryableError(err)
+		}
+
+		if contains(obj.statusReadyValue, fmt.Sprint(readyValue)) {
+			/* Resource is ready and we can exit the retry loop */
+			return nil
+		} else if len(obj.statusErrorValue) > 0 && contains(obj.statusErrorValue, fmt.Sprint(readyValue)) {
+			data, _ := json.Marshal(obj.apiData)
+			return resource.NonRetryableError(fmt.Errorf("restapi status got an error value [%s],details:\n %s", fmt.Sprint(readyValue), data))
+		} else {
+			return resource.RetryableError(fmt.Errorf("resource not yet ready - current value: %s", readyValue))
+		}
+	})
+	return err
+}
 func resourceRestAPIUpdate(d *schema.ResourceData, meta interface{}) error {
 	obj, err := makeAPIObject(d, meta)
 	if err != nil {
 		return err
 	}
-
+	// do nothong
+	if obj.updatePolicy == "none" {
+		return nil
+	} else if obj.updatePolicy == "create" {
+		return resourceRestAPICreate(d, meta)
+	} else if obj.updatePolicy == "recreate" {
+		err = resourceRestAPIDelete(d, meta)
+		if err != nil {
+			return err
+		}
+		return resourceRestAPICreate(d, meta)
+	}
 	/* If copy_keys is not empty, we have to grab the latest
 	   data so we can copy anything needed before the update */
 	client := meta.(*APIClient)
@@ -302,6 +338,10 @@ func resourceRestAPIUpdate(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("resource_api_object.go: Update routine called. Object built:\n%s\n", obj.toString())
 
 	err = obj.updateObject()
+	if err != nil {
+		return err
+	}
+	err = waitStatus(d, meta, obj)
 	if err == nil {
 		setResourceState(obj, d)
 	}
@@ -314,7 +354,10 @@ func resourceRestAPIDelete(d *schema.ResourceData, meta interface{}) error {
 		return err
 	}
 	log.Printf("resource_api_object.go: Delete routine called. Object built:\n%s\n", obj.toString())
-
+	if obj.deletePolicy != "delete" {
+		log.Print("resource_api_object.go: Delete is called, but the object is reserved")
+		return nil
+	}
 	err = obj.deleteObject()
 	if err != nil {
 		if strings.Contains(err.Error(), "404") {
@@ -417,9 +460,21 @@ func buildAPIObjectOpts(d *schema.ResourceData) (*apiObjectOpts, error) {
 	if v, ok := d.GetOk("query_string"); ok {
 		opts.queryString = v.(string)
 	}
-
-	opts.createReadyKey = d.Get("create_ready_key").(string)
-	opts.createReadyValue = d.Get("create_ready_value").(string)
+	if v, ok := d.GetOk("status_key"); ok {
+		opts.statusKey = v.(string)
+	}
+	if v, ok := d.GetOk("status_ready_value"); ok {
+		opts.statusReadyValue = toStringArray(v.([]interface{}))
+	}
+	if v, ok := d.GetOk("status_error_value"); ok {
+		opts.statusErrorValue = toStringArray(v.([]interface{}))
+	}
+	if v, ok := d.GetOk("delete_policy"); ok {
+		opts.deletePolicy = v.(string)
+	}
+	if v, ok := d.GetOk("update_policy"); ok {
+		opts.updatePolicy = v.(string)
+	}
 
 	readSearch := expandReadSearch(d.Get("read_search").(map[string]interface{}))
 	opts.readSearch = readSearch
